backend/requirements.txt
websockets
pytesseract
flask[async]
opencv-python
pdf2image
python-poppler
pillow
asyncio
langchain
